feat: move rewards balance and history to own hook file
